Move assertEqualStreams to TestUtil.
